+++ __tests__/plugin.test.js
@@ -1,4 +1,3 @@
-/* eslint no-template-curly-in-string: 0 */
 const configModule = require('../config');
 
 jest.mock("../config.js");
@@ -17,7 +16,7 @@ global.Cypress = {
 global.cy = {};
 
 describe('plugin', () => {
-  describe('initPlugin', () => {
+  it('initPlugin', () => {
     const globalConfig = {
       env: {
         "cypress-plugin-snapshots": {
@@ -33,92 +32,5 @@ describe('plugin', () => {
 
     initPlugin(on, globalConfig);
     expect(on).toBeCalledTimes(1);
-  });
-
-  describe('keepKeysFromExpected', () => {
-    const expected = {
-      field: "value",
-      items: [
-        {name: "item1", "items": [{name: "subitem1"}]},
-        {name: "item2"},
-        {name: "item3"},
-      ]
-    };
-
-    const actual = {
-      field: "value",
-      extra: "extra",
-      items: [
-        {name: "item1", "items": [{name: "subitem1", extra: 'extra1'}, {name: "subitem2", extra: 'extra2'}]},
-        {name: "item2"},
-        {name: "item3"},
-        {name: "item4"},
-      ]
-    };
-
-    it('ignoreExtraFields: false, ignoreExtraArrayItems: false', () => {
-      const config = {
-        ignoreExtraFields: false,
-        ignoreExtraArrayItems: false,
-      };
-      jest.spyOn(configModule, 'getConfig').mockImplementation(() => config);
-
-      const { keepKeysFromExpected } = require('../plugin');
-
-      const result = keepKeysFromExpected(actual, expected);
-      expect(result).toMatchSnapshot();
-    });
-
-    it('ignoreExtraFields: true, ignoreExtraArrayItems: false', () => {
-      const config = {
-        ignoreExtraFields: true,
-        ignoreExtraArrayItems: false,
-      };
-      jest.spyOn(configModule, 'getConfig').mockImplementation(() => config);
-
-      const { keepKeysFromExpected } = require('../plugin');
-      const result = keepKeysFromExpected(actual, expected);
-      expect(result).toMatchSnapshot();
-    });
-
-    it('ignoreExtraFields: true, ignoreExtraArrayItems: true', () => {
-      const config = {
-        ignoreExtraFields: true,
-        ignoreExtraArrayItems: true,
-      };
-      jest.spyOn(configModule, 'getConfig').mockImplementation(() => config);
-      const { keepKeysFromExpected } = require('../plugin');
-
-      const result = keepKeysFromExpected(actual, expected);
-      expect(result).toMatchSnapshot();
-    });
-
-    it('ignoreExtraFields: false, ignoreExtraArrayItems: true', () => {
-      const config = {
-        ignoreExtraFields: false,
-        ignoreExtraArrayItems: true,
-      };
-      jest.spyOn(configModule, 'getConfig').mockImplementation(() => config);
-
-      const { keepKeysFromExpected } = require('../plugin');
-
-      const result = keepKeysFromExpected(actual, expected);
-      expect(result).toMatchSnapshot();
-    });
-  });
-
-  describe('applyReplace', () => {
-    const { applyReplace } = require('../plugin');
-
-    it('object replace', () => {
-      const expected = {
-        "text": "Hi ${name} & ${name}",
-        "text2": "Hi ${name}",
-      };
-      const result = applyReplace(expected, {
-        name: 'Meinaart',
-      });
-      expect(result).toMatchSnapshot();
-    });
   });
 });