+++ config.js
@@ -1,5 +1,4 @@
 const crypto = require('crypto');
-const path = require('path');
 const { merge, cloneDeep } = require('lodash');
 
 function createToken() {
@@ -25,23 +24,11 @@ const CONFIG_KEY = 'cypress-plugin-snapshots';
 
 let config = DEFAULT_CONFIG;
 
-function resolveModulePath(filename) {
-  const fullPath = require.resolve(filename);
-  const parentDir = path.dirname(__dirname);
-  return path.join(
-    'node_modules/',
-    path.relative(parentDir, fullPath),
-  );
-}
-
 function initConfig(initialConfig) {
   config = cloneDeep(DEFAULT_CONFIG);
   if (initialConfig) {
     merge(config, initialConfig);
   }
-  config.DIFF_CSS_PATH = resolveModulePath('diff2html/dist/diff2html.css');
-  config.DIFF_JS_PATH = resolveModulePath('diff2html/dist/diff2html.js');
-  config.SOCKET_JS_PATH = resolveModulePath('socket.io-client/dist/socket.io.js');
   return config;
 }