+++ constants.js
@@ -1,9 +1,0 @@
-const MATCH = 'cypress-plugin-snapshot:match';
-const SAVE = 'cypress-plugin-snapshot:save';
-const URL_PREFIX = '#snapshot-compare-';
-
-module.exports = {
-  MATCH,
-  SAVE,
-  URL_PREFIX,
-};