+++ tasks/paths.js
@@ -1,0 +1,7 @@
+module.exports = {
+  PATH_CSS: './assets/styles.css',
+  PATH_DIFF_CSS: 'diff2html/dist/diff2html.css',
+  PATH_DIFF_JS: 'diff2html/dist/diff2html.js',
+  PATH_JS: './assets/script.js',
+  PATH_SOCKET_JS: 'socket.io-client/dist/socket.io.js',
+};