+++ tasks/task-names.js
@@ -1,0 +1,5 @@
+module.exports = {
+  GET_FILE: 'cypress-plugin-snapshot:getFile',
+  MATCH: 'cypress-plugin-snapshot:match',
+  SAVE: 'cypress-plugin-snapshot:save',
+};